Moving ai_modal to the right place in the directories structure
